jwt: Implement a better deterministic ECDSA
The current always uses the same nonce for signatures, which reveleals
enough information to decode the private key, after generating only two
signatures.

Since, we don't currently have support for a proper RNG, borrow the
determinsitic nonce generation from ed25519, namely, see using a hash of
the private key, and the message.  This still feeds the cprng_hmac,
which is a bit of extra work, but should still have similar security
properties.

This will use the same nonce, but only to sign exactly the same message
with the same private key, which doesn't reveal any additional
information.

See https://www.bbc.co.uk/news/technology-12116051 on why this is
important.